+++ app/src/androidTest/java/nextstep/payments/screen/card/list/RegisteredCreditCardsScreenTest.kt
@@ -12,7 +12,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class CreditCardScreenTest {
+class RegisteredCreditCardsScreenTest {
 
     @get:Rule
     val composeRule = createComposeRule()
