+++ app/src/main/java/nextstep/payments/ui/card/list/CardListViewModel.kt
@@ -4,19 +4,19 @@ import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import nextstep.payments.data.CreditCard
+import nextstep.payments.data.RegisteredCreditCards
 import nextstep.payments.data.PaymentCardsRepository
 
 class CardListViewModel : ViewModel() {
 
-    private val _creditCard = MutableStateFlow(CreditCard(emptyList()))
-    val creditCard: StateFlow<CreditCard> = _creditCard.asStateFlow()
+    private val _Registered_creditCards = MutableStateFlow(RegisteredCreditCards(emptyList()))
+    val registeredCreditCards: StateFlow<RegisteredCreditCards> = _Registered_creditCards.asStateFlow()
 
     init {
         fetchCards()
     }
 
     fun fetchCards() {
-        _creditCard.value = CreditCard(PaymentCardsRepository.cards)
+        _Registered_creditCards.value = RegisteredCreditCards(PaymentCardsRepository.cards)
     }
 }
